+++ src/main/java/nl/d2n/model/InsideBuildingResourceCost.java
@@ -5,14 +5,14 @@ import com.google.gson.annotations.Expose;
 public class InsideBuildingResourceCost {
 
     @Expose
-    private Integer itemId;
+    private Long itemId;
 
     @Expose
     private Integer amount;
 
-    public Integer getItemId() { return this.itemId; }
+    public Long getItemId() { return this.itemId; }
     public Integer getAmount() { return this.amount; }
 
-    public void setItemId(Integer itemId) { this.itemId = itemId; }
+    public void setItemId(Long itemId) { this.itemId = itemId; }
     public void setAmount(Integer amount) { this.amount = amount; }
 }
